working on player movement

+++ include/Player.hpp
@@ -8,14 +8,14 @@
 #include <Term.hpp>
 
 class Player : public Drawable, public KeyboardListener {
-  float x_pos, y_pos;
+  double x_pos, y_pos;
 
-  float deltaX, deltaY;
+  double deltaX, deltaY;
 
   ::std::shared_ptr<ts::Term> term;
   std::shared_ptr<EntityManager> entityManager;
 
-  void AddToDelta(float* delta, float add);
+  void AddToDelta(double* delta, double add);
 
   void FireBullet();
 
@@ -26,8 +26,8 @@ class Player : public Drawable, public KeyboardListener {
   void Draw();
   void Update();
 
-  float GetX();
-  float GetY();
+  double GetX();
+  double GetY();
 
   void SetX(uint32_t);
   void SetY(uint32_t);
